app/requirements.txt
fastapi
uvicorn
transformers==4.28.1  # Specify a compatible version
cassandra-driver
redis
tensorflow==2.9.1  # Specify a compatible version
pytest
tf-keras
